pyproject.toml
[tool.poetry]
name = "drf-api-checker"
version = "0.10.0"
description = ""
authors = ["sax <s.apostolico@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
djangorestframework = "^3.10"
pytz = "^2019.3"
django = "*"

[tool.poetry.dev-dependencies]
django-webtest = "^1.9"
factory-boy = "^2.12"
pyfakefs = "^3.7"
pytest-cov = "^2.8"
pytest-django = "^3.7"
pytest-echo = "^1.6"
pytest-pythonpath = "^0.7.3"
pdbpp = "^0.10.2"
sphinx = "^3.2.1"

[build-system]
requires = ["poetry"]
build-backend = "poetry.core.masonry.api"
